refactor: extract AdminPage into 8 components
Split the 1692-line AdminPage.tsx monolith into focused components
under web/src/components/admin/:

- ConfirmModal — with body scroll lock on mount
- DetailModal — with focus trap via useRef
- StatCard — reusable stat display card
- OverviewTab — stats, purge, vacuum controls
- ListingsTab — paginated listings management
- SearchesTab — search management with detail view
- UsersTab — user management with AnimatePresence fix
- LogsTab — log viewer with useMemo-optimized filtering

AdminPage.tsx is now 204 lines — tab orchestration and auth only.

Closes #582
