+++ web/src/components/admin/StatCard.tsx
@@ -1,0 +1,43 @@
+import type { Car } from "lucide-react";
+import { motion } from "motion/react";
+import { cn } from "@/lib/utils";
+
+export function StatCard({
+  label,
+  value,
+  icon: Icon,
+  color,
+  subtitle,
+}: {
+  label: string;
+  value: string | number;
+  icon: typeof Car;
+  color: string;
+  subtitle?: string;
+}) {
+  return (
+    <motion.div
+      initial={{ opacity: 0, y: 12 }}
+      animate={{ opacity: 1, y: 0 }}
+      className="rounded-2xl border border-border/50 bg-card p-5 card-hover"
+    >
+      <div className="flex items-center justify-between mb-3">
+        <span className="text-xs font-medium text-muted-foreground uppercase tracking-wider">
+          {label}
+        </span>
+        <div
+          className={cn(
+            "flex h-9 w-9 items-center justify-center rounded-xl",
+            color,
+          )}
+        >
+          <Icon className="h-[18px] w-[18px]" />
+        </div>
+      </div>
+      <p className="text-3xl font-bold tabular-nums text-foreground">{value}</p>
+      {subtitle && (
+        <p className="text-xs text-muted-foreground mt-1">{subtitle}</p>
+      )}
+    </motion.div>
+  );
+}
